Created theme system for hellwal
Modified the wallpaper script so that it reads a .hellwal file in he same directory as the image file.
This allows for the creation of custom theming using custom colors for each image file.
I also created a theme for each wallpaper I had using their default colors, to be modified later.